+++ packages/dumi-theme-sphinx-mdx/src/exports/DocumentRenderer.tsx
@@ -2,7 +2,6 @@ import { MDXProvider } from '@mdx-js/react';
 import { Alert } from 'antd';
 import { Link } from 'dumi';
 import Badge from 'dumi/theme-default/builtins/Badge';
-import Container from 'dumi/theme-default/builtins/Container';
 import SourceCode from 'dumi/theme-default/builtins/SourceCode';
 import Table from 'dumi/theme-default/builtins/Table';
 import React, { forwardRef } from 'react';
@@ -50,7 +49,6 @@ export const DocumentRenderer = forwardRef<HTMLElement, React.PropsWithChildren>
             components={{
               Link,
               Badge: Badge as unknown as ForeignComponent,
-              Container: Container as unknown as ForeignComponent,
               SourceCode: SourceCode as unknown as ForeignComponent,
               Table: Table as unknown as ForeignComponent,
               ...intrinsic,