+++ packages/dumi-theme-sphinx-mdx/src/internals/components/Target.tsx
@@ -1,0 +1,6 @@
+import { useFragmentFocus } from '../common/positioning.js';
+
+export const Target = ({ id, ...props }: React.ComponentProps<'span'>) => {
+  useFragmentFocus(id);
+  return <span id={id} {...props} />;
+};