+++ packages/unified-toolkit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @secretflow/unified-toolkit
 
+## 0.3.0
+
+### Minor Changes
+
+- fix issues with multiple element IDs
+
+## 0.2.1
+
+### Patch Changes
+
+- bb42f04: move dumi-plugin-search to alpha
+
 ## 0.2.1-alpha.0
 
 ### Patch Changes